Add playback progressbar

+++ app/src/main/java/me/timpushkin/voicenotes/models/Recording.kt
@@ -2,4 +2,4 @@ package me.timpushkin.voicenotes.models
 
 import android.net.Uri
 
-data class Recording(val name: String, val date: Long, val duration: Long, val uri: Uri)
+data class Recording(val name: String, val date: Long, val duration: Int, val uri: Uri)
